+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Called Numbers RESTful API route added (`http://x.x.x.x:8082/numbers`)
+- Called Number RESTful API resource live at (`http://x.x.x.x:8082/numbers/{number}`)
+- Called Numbers RESTful API resource live at (`http://x.x.x.x:8082/numbers`)
 
 ## [0.0.6] - 2022-01-21
 - Fix rpm build on Fedora due to wrongly defined libcurl callback definition issue causing `make check` to fail